Make "scroll down" notice prettier, show comment count
The scroll down notice now has a downward arrow and is no longer a well. The
comment count is shown next to each item, but it does not function as a link.
The JS now makes sure there is a 300 ms gap between AJAX completions before
showing the main view. This cuts down on the sequential loading, as it looks
like everything loads at once.