Add rudimentary comment support
Adds functionality to grab the top level comments of a post item, which are
individually grabbed and appended to the comment area when they finish loading.

Ideally I would like to wait for all of them to finish loading before
rendering them, but I've been having trouble doing so with an array of Deferred
Objects. For now the functionality works, regardless of whether it's perfectly
designed.

I haven't decided how to handle comment replies, and how deep of replies to go
to, but for now comments work.

All references to 'post' have also been replaced to 'item'.