Fix bug when there is one comment.
When there is one comment, the type format goes from arrays of arrays to a
single array, with the resolved object at the first index. This broke map,
since it was expecting an array, and got an object. It now checks to see if
there is only one comment and grabs the resolved object if there is.

This also fixes the error when there are no comments and the function tries to
run map() on an undefined value.